Improve static data fecthing
