+++ pages/_lang/generate.vue
@@ -144,7 +144,7 @@
 <script lang="ts">
 import Vue from "vue";
 import Component from "nuxt-class-component";
-import { Action, State, namespace } from "vuex-class";
+import { Action, State, Getter, namespace } from "vuex-class";
 
 import GeneratorMenu from "~/components/GeneratorMenu";
 import TwoWays from "~/components/TwoWays";
@@ -153,6 +153,7 @@ import * as generator from "~/store/modules/generator";
 
 const GeneratorState = namespace(generator.name, State);
 const GeneratorAction = namespace(generator.name, Action);
+const GeneratorGetter = namespace(generator.name, Getter);
 
 @Component({
   components: {
@@ -164,6 +165,8 @@ export default class extends Vue {
   public manifest$: generator.Manifest | null = null;
 
   @GeneratorState manifest: generator.Manifest;
+  @GeneratorGetter languagesNames: string[];
+  @GeneratorGetter displaysNames: string[];
 
   public created(): void {
     if (!this.manifest) {
